chore: extract URL/email utils to shared/lib/url-utils
getValidUrl, isWebUrl, addUrlProtocolPrefix, isValidEmail, and
isWebOrigin are pure utilities used by both UI and background.
Moving them to shared/lib/url-utils.ts lets UI import them without
crossing the architectural boundary; app/scripts/lib/util keeps
re-exporting them so background callers stay unaffected.
