+++ ui/components/app/assets/nfts/nft-details/nft-full-image.tsx
@@ -25,9 +25,7 @@ import {
 import { useI18nContext } from '../../../../../hooks/useI18nContext';
 import useGetAssetImageUrl from '../../../../../hooks/useGetAssetImageUrl';
 import useFetchNftDetailsFromTokenURI from '../../../../../hooks/useFetchNftDetailsFromTokenURI';
-// TODO: Remove restricted import
-// eslint-disable-next-line import-x/no-restricted-paths
-import { isWebUrl } from '../../../../../../app/scripts/lib/util';
+import { isWebUrl } from '../../../../../../shared/lib/url-utils';
 import { getNetworkConfigurationsByChainId } from '../../../../../../shared/lib/selectors/networks';
 import { getImageForChainId } from '../../../../../selectors/multichain';
 import {
