+++ ui/components/app/assets/nfts/nft-grid/nft-grid.tsx
@@ -10,9 +10,7 @@ import useGetAssetImageUrl from '../../../../../hooks/useGetAssetImageUrl';
 import { getImageForChainId } from '../../../../../selectors/multichain';
 import { getNetworkConfigurationsByChainId } from '../../../../../../shared/lib/selectors/networks';
 import useFetchNftDetailsFromTokenURI from '../../../../../hooks/useFetchNftDetailsFromTokenURI';
-// TODO: Remove restricted import
-// eslint-disable-next-line import-x/no-restricted-paths
-import { isWebUrl } from '../../../../../../app/scripts/lib/util';
+import { isWebUrl } from '../../../../../../shared/lib/url-utils';
 import {
   VirtualizedList,
   noAdjustmentsScroll,
