+++ ui/hooks/useIsOriginalNativeTokenSymbol.ts
@@ -5,9 +5,7 @@ import { getUseSafeChainsListValidation } from '../selectors';
 import { getMultichainCurrentNetwork } from '../selectors/multichain';
 import { isEvmChainId } from '../../shared/lib/asset-utils';
 
-// TODO: Remove restricted import
-// eslint-disable-next-line import-x/no-restricted-paths
-import { getValidUrl } from '../../app/scripts/lib/util';
+import { getValidUrl } from '../../shared/lib/url-utils';
 import { isOriginalNativeTokenSymbol } from '../helpers/utils/isOriginalNativeTokenSymbol';
 
 export function useIsOriginalNativeTokenSymbol(
