+++ ui/pages/onboarding-flow/privacy-settings/privacy-settings.tsx
@@ -20,8 +20,7 @@ import {
   FontWeight,
   TextButton,
 } from '@metamask/design-system-react';
-// eslint-disable-next-line import-x/no-restricted-paths
-import { addUrlProtocolPrefix } from '../../../../app/scripts/lib/util';
+import { addUrlProtocolPrefix } from '../../../../shared/lib/url-utils';
 import { TextField } from '../../../components/component-library';
 import {
   MetaMetricsEventCategory,
